+++ cryptography/lib/browser.dart
@@ -17,7 +17,7 @@
   'This library will be removed in a future major version.'
   ' You can find `BrowserCryptography` class in "package:cryptography/cryptography.dart".',
 )
-library cryptography.browser;
+library;
 
 export 'src/browser/browser_cryptography_when_not_browser.dart'
-    if (dart.library.html) 'src/browser/browser_cryptography.dart';
+    if (dart.library.js_interop) 'src/browser/browser_cryptography.dart';