+++ cryptography/test/browser_cryptography_test.dart
@@ -12,7 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-@TestOn('chrome')
+@TestOn('browser')
+library;
+
 import 'package:cryptography/cryptography.dart';
 import 'package:cryptography/dart.dart';
 import 'package:cryptography/src/browser/aes_cbc.dart';